reorientació del sistema de nous cursos i importació

+++ PIC/ViewModel/FinalitzarCursVM.cs
@@ -1,14 +1,17 @@
-﻿using PIC.APIClient;
+﻿using Microsoft.Win32;
+using PIC.APIClient;
 using PIC.Model;
 using PIC.Utilities;
 using PIC.View;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Input;
+using System.Globalization;
 
 namespace PIC.ViewModel
 {
@@ -20,6 +23,7 @@ namespace PIC.ViewModel
         private readonly UsuarisApiClient _usuarisApiClient;
         private readonly CursosApiClient _cursosApiClient;
         private readonly RegistresApiClient _registresApiClient;
+        private readonly DispositiusApiClient _dispositiusApiClient;
         public MissatgeErrorVM MissatgeError { get; set; }
         public NotificacioVM Notificacio { get; set; }
 
@@ -37,6 +41,7 @@ namespace PIC.ViewModel
             _usuarisApiClient = new UsuarisApiClient();
             _cursosApiClient = new CursosApiClient();
             _registresApiClient = new RegistresApiClient();
+            _dispositiusApiClient = new DispositiusApiClient();
         }
 
         // CODI DE SEGURETAT
@@ -75,6 +80,22 @@ namespace PIC.ViewModel
             }
         }
 
+        // RUTA DE FITXER
+
+        private string _rutaFitxer = "";
+        public string RutaFitxer
+        {
+            get => _rutaFitxer;
+            set { _rutaFitxer = value; OnPropertyChanged(); }
+        }
+
+        private string _botoRutaFitxer = "[...]";
+        public string BotoRutaFitxer
+        {
+            get => _botoRutaFitxer;
+            set { _botoRutaFitxer = value; OnPropertyChanged(); }
+        }
+
         // VISIBILITAT MENU
         private Visibility _esVisble = Visibility.Collapsed;
         public Visibility EsVisible
@@ -103,19 +124,51 @@ namespace PIC.ViewModel
             EsVisible = Visibility.Collapsed;
         });
 
+        // SELECCIONAR FITXER
+        public ICommand SeleccionarFitxer_Click => new RelayCommand(o =>
+        {
+            OpenFileDialog openFileDialog = new OpenFileDialog();
+            openFileDialog.Filter = "Fitxers Excel (*.xlsx)|*.xlsx|Fitxers CSV (*.csv)|*.csv";
+
+            if (openFileDialog.ShowDialog() == true)
+            {
+                RutaFitxer = openFileDialog.FileName;
+
+                // Nom de l'arxiu per al botó
+                string[] rutaSeparada = RutaFitxer.Split('\\');
+                BotoRutaFitxer = rutaSeparada[rutaSeparada.Count() - 1];
+            }
+        });
+
         // FINALITZAR CURS
         public ICommand FinalitzarCurs_Click => new RelayCommand(async _ =>
         {
             // Si es pot finalitzar
-            if (!esPotFinalitzar) 
-            { 
-                return;
-            }
-            
+            //if (!esPotFinalitzar) 
+            //{ 
+            //    return;
+            //}
+
+            //esPotFinalitzar = false;
+
+            //using (var reader = new StreamReader(RutaFitxer))
+            //using (var csv = new CsvReader(reader, CultureInfo.InvariantCulture))
+            //{
+            //    // Si tens una classe que coincideix amb les columnes
+            //    var records = csv.GetRecords<Usuari>().ToList();
+
+            //    foreach (var usuari in records)
+            //    {
+            //        Console.WriteLine($"{usuari.Nom} - {usuari.Email}");
+            //    }
+            //}
+
+            return;
             // Si el codi de seguretat NO coincideix
             if (CodiSeguretat.ToString() != CodiSeguretatTextBox)
             {
                 MissatgeError.Mostrar("El codi de seguretat no coincideix.");
+                esPotFinalitzar = true;
                 return;
             }
                            
@@ -129,19 +182,96 @@ namespace PIC.ViewModel
             if (prestecs == null)
             {
                 MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els préstecs. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                esPotFinalitzar = true;
+                return;
+            }
+
+            var usuaris = await _usuarisApiClient.GetAllUsuarisAsync();
+
+            // Si la consulta falla
+            if (usuaris == null)
+            {
+                MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els usuaris. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                esPotFinalitzar = true;
                 return;
             }
 
             foreach (var prestec in prestecs)
             {
-                TextProces = $"// Esborrant préstec [{prestec.Id}]...";
-                var prestecEsborrat = await _prestecsApiClient.DeletePrestecAsync(prestec.Id);
-
-                if (prestecEsborrat == -1)
+                foreach (var usuari  in usuaris)
                 {
-                    MissatgeError.Mostrar($"Hi ha hagut un problema al esborrar el préstec amb ID[{prestec.Id}]. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
-                    return;
-                }
+                    // Si el préstec és d'un alumne
+                    if (usuari.Tipus.ToLower() == "Alumne")
+                    {
+                        TextProces = $"// Esborrant préstec [{prestec.Id}]...";
+                        var prestecEsborrat = await _prestecsApiClient.DeletePrestecAsync(prestec.Id);
+
+                        Dispositiu dispositiuConsultaFinalitzar = await _dispositiusApiClient.GetDispositiuPerIdAsync(prestec.IdDispositiu);
+
+                        // Si la consulta falla
+                        if (dispositiuConsultaFinalitzar == null)
+                        {
+                            MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els dispositius. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                            esPotFinalitzar = true;
+                            return;
+                        }
+
+                        Dispositiu dispositiuPrestatFinalitzar = new Dispositiu();
+                        dispositiuPrestatFinalitzar.Id = prestec.IdDispositiu;
+                        dispositiuPrestatFinalitzar.Nom = dispositiuConsultaFinalitzar.Nom;
+                        dispositiuPrestatFinalitzar.IdCategoria = dispositiuConsultaFinalitzar.IdCategoria;
+                        dispositiuPrestatFinalitzar.Estat = "Disponible";
+
+                        int confirmarDispositiuPrestatFinalitzar = await _dispositiusApiClient.UpdateDispositiuAsync(dispositiuPrestatFinalitzar);
+
+                        // Si la consulta falla
+                        if (confirmarDispositiuPrestatFinalitzar == -1)
+                        {
+                            MissatgeError.Mostrar($"Hi ha hagut un problema al actualitzar el dispositius. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                            esPotFinalitzar = true;
+                            return;
+                        }
+
+                        Usuari usuariConsultaFinalitzar = await _usuarisApiClient.GetUsuariPerIdAsync(prestec.IdUsuari);
+
+                        // Si la consulta falla
+                        if (usuariConsultaFinalitzar == null)
+                        {
+                            MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els usuaris. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                            esPotFinalitzar = true;
+                            return;
+                        }
+
+                        // Crear registre
+                        Registre registreFinalitzat = new Registre();
+                        registreFinalitzat.IdPrestec = prestec.Id;
+                        registreFinalitzat.Accio = "No finalitzat";
+                        registreFinalitzat.NomUsuari = $"{usuariConsultaFinalitzar.Nom} {usuariConsultaFinalitzar.Cognom}";
+                        registreFinalitzat.IdUsuari = (int)usuariConsultaFinalitzar.Id;
+                        registreFinalitzat.NomDispositiu = dispositiuConsultaFinalitzar.Nom;
+                        registreFinalitzat.IdDispositiu = (int)dispositiuConsultaFinalitzar.Id;
+                        registreFinalitzat.NomGrup = usuariConsultaFinalitzar.Grup;
+                        registreFinalitzat.IdGrup = (int)usuariConsultaFinalitzar.IdGrup;
+                        registreFinalitzat.DataAccio = DateTime.Now;
+                        registreFinalitzat.DataRetorn = prestec.DataRetorn;
+
+                        Registre registreCreatFinalitzar = await _registresApiClient.PostRegistreAsync(registreFinalitzat);
+
+                        if (registreCreatFinalitzar == null)
+                        {
+                            MissatgeError.Mostrar($"Hi ha hagut un problema al crear el registre. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                            esPotFinalitzar = true;
+                            return;
+                        }
+
+                        if (prestecEsborrat == -1)
+                        {
+                            MissatgeError.Mostrar($"Hi ha hagut un problema al esborrar el préstec amb ID[{prestec.Id}]. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                            esPotFinalitzar = true;
+                            return;
+                        }
+                    }
+                }                
             }
 
             // Esborrar alumnes
@@ -152,6 +282,7 @@ namespace PIC.ViewModel
             if (alumnes == null)
             {
                 MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els alumnes. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                esPotFinalitzar = true;
                 return;
             }
 
@@ -166,11 +297,13 @@ namespace PIC.ViewModel
                 if (alumneEsborrat == -1)
                 {
                     MissatgeError.Mostrar($"Hi ha hagut un problema al esborrar l'alumne amb ID[{alumne.IdUsuari}]. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                    esPotFinalitzar = true;
                     return;
                 }
                 if (usuariEsborrat == -1)
                 {
                     MissatgeError.Mostrar($"Hi ha hagut un problema al esborrar l'usuari amb ID[{alumne.IdUsuari}]. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                    esPotFinalitzar = true;
                     return;
                 }
             }
@@ -183,6 +316,8 @@ namespace PIC.ViewModel
             if (cursos == null)
             {
                 MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els cursos. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                esPotFinalitzar = true;
+                return;
             }
 
             foreach (var curs in cursos)
@@ -193,29 +328,7 @@ namespace PIC.ViewModel
                 if (alumneEsborrat == -1)
                 {
                     MissatgeError.Mostrar($"Hi ha hagut un problema al esborrar el curs amb ID[{curs.Id}]. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
-                    return;
-                }
-            }
-
-            // Esborrar registres
-            TextProces = "// Consultant registres...";
-            var registres = await _registresApiClient.GetAllRegistresAsync();
-
-            // Si la consulta falla
-            if (registres == null)
-            {
-                MissatgeError.Mostrar($"Hi ha hagut un problema al consultar els registres. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
-                return;
-            }
-
-            foreach (var registre in registres)
-            {
-                TextProces = $"// Esborrant el registre amb ID[{registre.Id}]...";
-                var alumneEsborrat = await _registresApiClient.DeleteRegistreAsync((int)registre.Id);
-
-                if (alumneEsborrat == -1)
-                {
-                    MissatgeError.Mostrar($"Hi ha hagut un problema al esborrar el registre [{registre.Id}]. Revisa la comunicació entre la API i l'aplicació i torna a executar aquesta funció.");
+                    esPotFinalitzar = true;
                     return;
                 }
             }
